pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "milestone1"
version = "0.1.0"
description = "AI-powered restaurant recommendation system (Zomato use case)"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.18",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13",
    "typer>=0.12",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
]
streamlit = [
    "streamlit>=1.35",
]
dev = [
    "pytest>=8",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
milestone1 = "milestone1.phase0.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that hit external services (deselect with -m 'not integration')",
]
